+++ lib/doorkeeper/oauth/refresh_token_request.rb
@@ -35,8 +35,10 @@ module Doorkeeper
         refresh_token.transaction do
           refresh_token.lock!
           raise Errors::InvalidTokenReuse if refresh_token.revoked?
+          if refresh_token.revoked_at.nil?
+            refresh_token.revoke_in(server.refresh_token_revoked_in)
+          end
 
-          refresh_token.revoke
           create_access_token
         end
       end
@@ -56,7 +58,9 @@ module Doorkeeper
           resource_owner_id: refresh_token.resource_owner_id,
           scopes: scopes.to_s,
           expires_in: expires_in,
-          use_refresh_token: true)
+          use_refresh_token: true,
+          previous_refresh_token: refresh_token.refresh_token.to_s
+        )
       end
 
       def validate_token_presence